decouple bounding-boxes
